Cleaned up quest add verification. Moved createQuest to QuestDB.
