+++ modules/control/Utils.js
@@ -1,99 +1,7 @@
-import QuestDB       from './QuestDB.js';
-import Socket        from './Socket.js';
-import ViewManager   from './ViewManager.js';
-import Quest         from '../model/Quest.js';
-import QuestFolder   from '../model/QuestFolder.js';
-
-import { constants, questTypes, questTypesI18n, settings } from '../model/constants.js';
+import { constants, settings } from '../model/constants.js';
 
 export default class Utils
 {
-   /**
-    * Creates a new quest and waits for the journal entry to update and QuestDB to pick up the new Quest which
-    * is returned.
-    *
-    * @param {object}   options - Optional parameters.
-    *
-    * @param {object}   data - Quest data to assign to new quest.
-    *
-    * @param {string}   parentId - Any associated parent ID; if set then this is a subquest.
-    *
-    * @param {boolean}  notify - Post a UI message.
-    *
-    * @returns {Promise<Quest>} The newly created quest.
-    */
-   static async createQuest({ data = {}, parentId = void 0, notify = false, swapTab = false } = {})
-   {
-      // Get the default permission setting and attempt to set it if found in ENTITY_PERMISSIONS.
-      const defaultPerm = game.settings.get(constants.moduleName, settings.defaultPermission);
-
-      const permission = {
-         default: typeof CONST.ENTITY_PERMISSIONS[defaultPerm] === 'number' ? CONST.ENTITY_PERMISSIONS[defaultPerm] :
-          CONST.ENTITY_PERMISSIONS.OBSERVER
-      };
-
-      const trustedPlayerEdit = Utils.isTrustedPlayer();
-
-      // Used for a player created quest setting and the quest as 'available' for normal players or 'hidden' for
-      // trusted players.
-      if (!game.user.isGM)
-      {
-         data.status = trustedPlayerEdit ? questTypes.inactive : questTypes.available;
-         permission[game.user.id] = CONST.ENTITY_PERMISSIONS.OWNER;
-      }
-
-      const parentQuest = QuestDB.getQuest(parentId);
-      if (parentQuest)
-      {
-         data.parent = parentId;
-      }
-
-      // Creating a new quest will add any missing data / schema.
-      const tempQuest = new Quest(data);
-
-      const entry = await JournalEntry.create({
-         name: tempQuest.name,
-         folder: QuestFolder.get().id,
-         permission,
-         flags: {
-            [constants.moduleName]: {
-               json: tempQuest.toJSON()
-            }
-         }
-      });
-
-      if (parentQuest)
-      {
-         parentQuest.addSubquest(entry.id);
-         await parentQuest.save();
-         Socket.refreshQuestPreview({ questId: parentQuest.id });
-      }
-
-      const quest = QuestDB.getQuest(entry.id);
-
-      if (notify)
-      {
-         ui.notifications.info(game.i18n.format('ForienQuestLog.Notifications.QuestAdded', {
-            name: quest.name,
-            status: game.i18n.localize(questTypesI18n[quest.status])
-         }));
-      }
-
-      if (swapTab)
-      {
-         const questLog = ViewManager.questLog;
-         if (questLog.rendered && questLog._tabs[0] && questLog._tabs[0].active !== quest.status)
-         {
-            questLog._tabs[0].activate(quest.status);
-         }
-      }
-
-      // Players don't see Hidden tab, but assistant GM can, so emit anyway
-      Socket.refreshAll();
-
-      return quest;
-   }
-
    /**
     * A convenience method to return the module data object for FQL.
     *
@@ -207,20 +115,6 @@ export default class Utils
    static isTrustedPlayer(user = game.user)
    {
       return user.isTrusted && game.settings.get(constants.moduleName, settings.trustedPlayerEdit);
-   }
-
-   /**
-    * Convenience method to determine if the QuestTracker is visible to the current user. Always for the GM when
-    * QuestTracker is enabled, but only for users if `hideFromPlayers` is false. There must also be active quests for
-    * the tracker to be visible.
-    *
-    * @returns {boolean} Whether the QuestTracker is visible.
-    */
-   static isQuestTrackerVisible()
-   {
-      return game.settings.get(constants.moduleName, settings.enableQuestTracker) &&
-       (game.user.isGM || !game.settings.get(constants.moduleName, settings.hideFQLFromPlayers)) &&
-        QuestDB.getActiveCount() > 0;
    }
 
    /**
